+++ src/model/invite.rs
@@ -236,7 +236,7 @@ impl InviteGuild {
     #[cfg(all(feature = "cache", feature = "utils"))]
     #[inline]
     #[must_use]
-    pub fn shard_id(&self, cache: impl AsRef<Cache>) -> u32 {
+    pub fn shard_id(&self, cache: impl AsRef<Cache>) -> u16 {
         self.id.shard_id(&cache)
     }
 
@@ -261,7 +261,7 @@ impl InviteGuild {
     #[cfg(all(feature = "utils", not(feature = "cache")))]
     #[inline]
     #[must_use]
-    pub fn shard_id(&self, shard_count: u32) -> u32 {
+    pub fn shard_id(&self, shard_count: u16) -> u16 {
         self.id.shard_id(shard_count)
     }
 }
